sched/deadline: avoid pointless __setscheduler()
There is no need to dequeue/enqueue and push/pull if there are no
scheduling parameters changed for the DL class.

Both fair and RT classes already check if parameters changed for them to
avoid unnecessary overhead. This patch add the parameters changed test
for the DL class in order to reduce overhead.

Signed-off-by: Wanpeng Li <wanpeng.li@linux.intel.com>
[ Fixed up the changelog. ]
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Cc: Juri Lelli <juri.lelli@arm.com>
Cc: Kirill Tkhai <ktkhai@parallels.com>
Cc: Linus Torvalds <torvalds@linux-foundation.org>
Link: http://lkml.kernel.org/r/1416962647-76792-5-git-send-email-wanpeng.li@linux.intel.com
Signed-off-by: Ingo Molnar <mingo@kernel.org>
Signed-off-by: Michele Beccalossi <beccalossi.michele@gmail.com>